Small tweaks to the coverage file
